Add endOnBlankLine as an option for the parser
Instead of always ending the parse on a blank line, offer it as an option.
This will be useful to users whose .csv file contains multiple CSV documents
with different columns etc.
The new default is to treat a blank line as a record containing one field
whose content is the empty string.

For an example of endOnBlankLine in action, see examples/Concat.hs