Generate benchmark data on-the-fly
This allows us to remove the massive csv files previously used for
benchmarking. They're clogging up the git history. The intention is
to use git filter-branch to remove them from history.

The benchmarks have suffered (this basically cuts the suite in half)
but that's a small price to pay and can be remedied later.